requirements.txt
paho-mqtt
loguru